refactor: adds 'mega' folder
