+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import type { AppProps } from 'next/app';
 import { useState } from 'react';
-import { ThemeContext, themeDefaultValue } from '../components/macro/Context/';
-import type { iThemeProps } from '../components/macro/Context/';
+import { ThemeContext, themeDefaultValue } from '../components/mega/Context';
+import type { iThemeProps } from '../components/mega/Context';
 import '../styles/global.scss';
 
 export default function MyApp({ Component, pageProps }: AppProps) {
